fix packet format and add examples

+++ src/server/config.zig
@@ -4,7 +4,7 @@ const CHALLENGE_KEY_SIZE = @import("../root.zig").CHALLENGE_KEY_SIZE;
 const SECRET_KEY_SIZE = @import("../root.zig").SECRET_KEY_SIZE;
 
 pub const ServerConfig = struct {
-    protocol_id: u64,
+    protocol_id: u32,
     handshake_alive_ns: u64,
     client_timeout_ns: u64,
     public_addresses: []const std.net.Address,
@@ -13,7 +13,7 @@ pub const ServerConfig = struct {
     secret_key: ?[SECRET_KEY_SIZE]u8,
 
     pub fn init(
-        protocol_id: u64,
+        protocol_id: u32,
         handshake_alive_ns: u64,
         client_timeout_ns: u64,
         public_addresses: []const std.net.Address,
